Update functions.php
In the_bootstrap_comment_form_defaults, the legend element is used as a child of an h3 element (presumably for styling). According to W3C, legend is only valid as the first child of a fieldset. http://www.w3.org/TR/html-markup/legend.html The HTML it throwing validator errors as a result.

Replace the legend element with span.legend so it can be styled.